fix: move selected from Groups to NewQuizForm
